+++ features/spatial_denoising/README.md
@@ -2,10 +2,16 @@
 
 项目：Spatail_Denoising
 
-去噪代码：clear_noise.py
-测试代码：test.py
-待处理图片：test.jpeg
-处理后的图片：处理结果.jpeg
+去噪代码：clear_noise.py     
+单元测试代码：test_1.py    
+待处理图片：test.jpeg 
+处理后的图片：处理结果.jpeg    
 预期图片：原图.bmp
+
+测试方法：运行test_1.py文件  
+
+存在问题:运行根目录test.sh进行测试时，需要将test.jepg和原图.bmp移动至根目录。     
+
+
 
 #----------------------------------#
